Allow switching between p95, duration, percentage
