+++ static/app/views/starfish/views/webServiceView/spanGroupBreakdownContainer.tsx
@@ -1,7 +1,9 @@
+import {useState} from 'react';
 import {useTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 
 import {getInterval} from 'sentry/components/charts/utils';
+import {SelectOption} from 'sentry/components/compactSelect';
 import {Panel} from 'sentry/components/panels';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
@@ -42,12 +44,28 @@ export type DataRow = {
   group: Group;
 };
 
+export enum DataDisplayType {
+  DURATION_P95 = 'duration_p95',
+  CUMULATIVE_DURATION = 'cumulative_duration',
+  PERCENTAGE = 'percentage',
+}
+
 export function SpanGroupBreakdownContainer({transaction, transactionMethod}: Props) {
   const pageFilter = usePageFilters();
   const organization = useOrganization();
   const location = useLocation();
   const {selection} = pageFilter;
   const theme = useTheme();
+
+  const options: SelectOption<DataDisplayType>[] = [
+    {label: 'Duration (p95)', value: DataDisplayType.DURATION_P95},
+    {label: 'Total Duration', value: DataDisplayType.CUMULATIVE_DURATION},
+    {label: 'Percentages', value: DataDisplayType.PERCENTAGE},
+  ];
+
+  const [dataDisplayType, setDataDisplayType] = useState<DataDisplayType>(
+    DataDisplayType.DURATION_P95
+  );
 
   const {data: segments, isLoading: isSegmentsLoading} = useDiscoverQuery({
     eventView: getCumulativeTimeEventView(
@@ -87,6 +105,7 @@ export function SpanGroupBreakdownContainer({transaction, transactionMethod}: Pr
         transactionMethod ? `http.method:${transactionMethod}` : ''
       }`,
       ['span.category'],
+      dataDisplayType,
       true
     ),
     enabled: true,
@@ -169,6 +188,9 @@ export function SpanGroupBreakdownContainer({transaction, transactionMethod}: Pr
         isCumulativeTimeLoading={isCumulativeDataLoading}
         transaction={transaction}
         errored={isError}
+        options={options}
+        dataDisplayType={dataDisplayType}
+        setDataDisplayType={setDataDisplayType}
       />
     </StyledPanel>
   );
@@ -183,12 +205,18 @@ const getEventView = (
   pageFilters: PageFilters,
   query: string,
   groups: string[],
+  dataDisplayType: DataDisplayType,
   getTimeseries?: boolean
 ) => {
+  const yAxis =
+    dataDisplayType === DataDisplayType.DURATION_P95
+      ? `p95(${SPAN_SELF_TIME})`
+      : `sum(${SPAN_SELF_TIME})`;
+
   return EventView.fromSavedQuery({
     name: '',
     fields: [`sum(${SPAN_SELF_TIME})`, `p95(${SPAN_SELF_TIME})`, ...groups],
-    yAxis: getTimeseries ? [`p95(${SPAN_SELF_TIME})`] : [],
+    yAxis: getTimeseries ? [yAxis] : [],
     query,
     dataset: DiscoverDatasets.SPANS_METRICS,
     start: pageFilters.datetime.start ?? undefined,
